+++ src/server/request.rs
@@ -9,7 +9,8 @@ use std::io::net::tcp::TcpStream;
 use {HttpResult};
 use version::{HttpVersion};
 use method;
-use header::{Headers, ContentLength};
+use header::Headers;
+use header::common::ContentLength;
 use rfc7230::{read_request_line};
 use rfc7230::{HttpReader, SizedReader, ChunkedReader};
 use uri::RequestUri;
