+++ src/server/response.rs
@@ -8,6 +8,7 @@ use std::io::net::tcp::TcpStream;
 use time::now_utc;
 
 use header;
+use header::common;
 use status;
 use version;
 use rfc7230::{CR, LF, LINE_ENDING};
@@ -48,8 +49,8 @@ impl Response {
         debug!("writing head: {} {}", self.version, self.status);
         try!(write!(self.body, "{} {}{}{}", self.version, self.status, CR as char, LF as char));
 
-        if !self.headers.has::<header::Date>() {
-            self.headers.set(header::Date(now_utc()));
+        if !self.headers.has::<common::Date>() {
+            self.headers.set(common::Date(now_utc()));
         }
 
         for (name, header) in self.headers.iter() {
